[TextFields] Change label animation again (#8703)
There's been this bug for a long time now with the floating label animation (only on simulators using below iOS 13, and not on devices, for some reason) where the animation is sometimes skipped entirely. I thought I had a fix for it with #8700, but when I started preparing the PR for the outlined textfields (which is blocked on #8674) I noticed that it was still a problem for them, for some reason. This PR I _think_ fixes it for everything though, despite being more complicated than I'd like.

Related to #6942.