[TextFields] Add outlined textfield (#8705)
This PR adds the outlined textfields. The snapshots look a little weird because the way the snapshot tests are currently implemented they only get an image that's as big as the frame of the view. In a follow up PR I'm going to change that, so the images will have the entire floating label (which goes above the top of the view) for example.

![textfield-example](https://user-images.githubusercontent.com/8020010/67978102-ba212280-fbef-11e9-8802-7a0b230f900a.gif)


Closes #6942.