[ActionSheet] Move layout code to `viewDidLayoutSubviews` (#8735)
This change from viewWillLayoutSubviews to viewDidLayoutSubviews allows the labels within the header to have the correct font when sizeThatFits is called. This fixes a bug where sizeThatFits wouldn't be correct then the ActionSheet table would be incorrect because it would think that the header was smaller/larger than it actually was.