[ThumbTrack] Discrete dots scale slowly with track height. (#8744)
The discrete dots in the ThumbTrack originally assumed that the track would
only ever be 2 points high. Now that customizable heights are allowed, the
dots should scale more slowly.

This is still not perfect, since the original positioning of the ticks resulted in the filled track stopping at non-center positions of each tick. However, it is an incremental improvement and better VX will necessitate Material Design inputs.

![Example showing misaligned track and tick mark.](https://user-images.githubusercontent.com/1753199/68304829-95f09600-0074-11ea-90c0-90b1a344327e.png)

Improves #8741