[Snackbar] Add will present block to allow customization of sna… (#8751)
Add a will present block to the MDCSnackbarMessage class. This allows for clients who wish to customize their snackbar message view (MDCSnackbarMessageView) instance, but without exposing UI properties on the message objects. This also makes it easier for clients who have multiple files where they show the snackbar with only one Manager to not have issues with different classes stealing the MDCSnackbarDelegate in order to add customization to the snackbar view.

![Simulator Screen Shot - iPhone 7 - 2019-11-07 at 12 38 42](https://user-images.githubusercontent.com/4066863/68414853-fcea7980-015e-11ea-8ae6-3ceb7d091a2e.png)