[AppBar] Add API to auto adjust height based on HeaderStackView for app bar. (#8764)
Part of https://github.com/material-components/material-components-ios/issues/8677.

This API helps to reduce the complexity of calculating minimumHeight and maximumHeight on FlexibleHeader when user wants to update the app bar height based on header stack view.

This flag is required to be set during configuration stage. The reset behavior (turning back to NO) was not implemented because it would require adding API to minMaxHeight class and all configuration needs be to be set before flexible header is used. This feature can be added if needed.