[Dragons] Simplify the MDCDragonsTableViewCell accessory types.
The accessory views do not need to be buttons. They are now image views and the names of the properties have been updated to reflect their functional purpose. From a public point of view, the properties only need to be UIViews, so the property types have been updated accordingly as well.

Polish as part of https://github.com/material-components/material-components-ios/pull/8856