[FlexibleHeader] Add FlexibleHeaderShiftedOffscreenWithShiftBehaviorDisabledTests.
These tests validate the behavior of the flexible header when the shift behavior is disabled.

Notably, these tests reveal a bug in the implementation of isShiftedOffscreen which can occur with the following repro steps:

1. Set the flexible header's shift behavior to disabled and that it has a tracking scroll view with a content size greater than its bounds by at least the height of the header (in order to allow the scroll view to be dragged).
2. Hide the flexible header with shiftHeaderOffScreenAnimated:
3. Drag the flexible header's tracking scroll view up and then down, revealing the flexible header.
4. Inspect the state of isShiftedOffscreen.

Expected value: false
Actual value: true

This is due to the fact that shiftHeaderOffScreenAnimated: sets an internal bit that indicates that the flexible header *wants* to be hidden, and this bit is not reset once the user starts interacting with the flexible header. The likely fix is to clear the "wants to be hidden" flag when the user starts interacting with the flexible header again.

Discovered as part of https://github.com/material-components/material-components-ios/issues/5185