Clean up unsupported situation when inferring the top safe area inset
In the case of automatically selecting a view controller, when the selected view controller is self.topLayoutGuideViewController, self.topLayoutGuideViewController.parentViewController is guaranteed to be nil, and [self fhv_rootAncestorOfViewController:nil] returns nil. Therefore, just set ancestor = nil directly.

An argument could be made that the same check for self.topLayoutGuideAdjustmentEnabled && ancestor == self.topLayoutGuideViewController should be made when the delegate returns a value, but that would result in a behavior change (infinitely increasing the top layout guide vs immediately crashing).

PiperOrigin-RevId: 283419359