[TextFields] Remove MDCTextFieldTypographyThemer from snapshot tests (#9196)
This goal of this PR is to remove usage of MDCTextFieldTypographyThemer from MDCTextField snapshot tests in preparation for its deprecation and deletion. In cases where theming with a theming extension/container scheme is available I'm changing the snapshot test files to use those. In cases where there is no equivalent theming extension I am foregoing theming. The two things that have themers but lack theming extensions are MDCTextField itself (we only have theming extensions for the filled and outlined textfield controllers) and the MDCTextInputControllerOutlinedTextArea.

Related to #9168