[TextFields] Remove usage of MDCOutlinedTextFieldColorThemer (#9214)
This goal of this PR is to remove usage of MDCTextFieldColorThemer in preparation for its deprecation and deletion. The snapshot tests that were using this themer no longer have color theming. There is a TODO in each of these tests saying to use a theming extension if/when one that will work for them is written.

Related to #9080