Revert "[scripts/release] Always branch off of stable. (#9310)" (#9341)
This reverts commit ca81e6e42826956e1912665a8e8077ab0d9705be.

The reverted commit does not actually fix the ancestry problem for a given release and the change was adding an unnecessary merge operation as a result. Simply branching off of `develop` is sufficient to create history from `stable` to `develop`.

In a follow-up change, I will address the actual root problem of `develop` not having an ancestral relations to the `stable` release commits.