[TextControls] Use CGFloat consts in theming extensions (#9360)
This PR makes it so that the opacity values in the TextControls theming extensions use CGFloat consts.

c497441 also resolves an issue related to MDCTextControl textfield disabled opacity values being higher than they should be. I noticed the disabled opacity looked off while working on #9358. This gif, which demonstrates the theming extension value changes in this PR, was taken from that PR:
![example](https://user-images.githubusercontent.com/8020010/71596471-8ab66580-2b0d-11ea-81c5-e1a775d8db14.gif)

Closes #9361.