[Docs] Replace releasing instructions with general steps.
The releasing process is being combined in a Google-internal document to
reduce errors introduced by jumping between the two documents.  The current
`releasing.md` is being replaced with general information about the release
process that may be relevant to clients and contributors.

The previous contents of this file have already been imported to Google by
cl/289097261.

Part of #9399