[Chips] Add snapshot tests for selectedImage layout.
Adds more snapshot tests to capture the behavior of Chips when a
`selectedImageView.image` is set. An upcoming change will correct the layout
of subviews when `imagePadding` is used, and these tests capture errors that
take place when a `selectedImageView.image` is set.

Preparation for #9250

This is a roll-forward of #9266