[ActionSheet] Action sheet divider (#9449)
This adds the ability for MDCActionSheet to add dividers on a per action basis. The dividers are always represented above the action and respect the `contentInsets` property of the `MDCActionSheetController`. 

Any update to the action before the `MDCActionSheetController` is presented will be displayed when presented. If an ActionSheetController is presented then this property is changed those changes will not be present. I've filed #9465 to track that work. I choose not to add this functionality since similar properties, (tintColor, titleColor, title, image) on MDCActionSheetAction also follow this pattern. 

Closes #9331