[FeatureHighlight] Allow UIAccessibility to activate the "dismiss" affordance.
This change connects the recently-added "dismiss" affordance to an action that
allows UIAccessibility to reject the Feature Highlight.  In #8959, a
discoverable affordance was added to Feature Highlight so UIAccessibility
would present dismiss options more easily to users. That change did not
include connecting the view to any action or gesture recognizer, so users
could discover a "dismiss" button but nothing happened when it was activated.

This change does the following:

*   Changes the dismiss view to a UIButton. There were complications when
    attempting to add another gesture recognizer to the view hierarchy and a
    UIButton was a simpler solution.
*   Sets the layout frame of the dismiss view to `{0, 0, 1, 1}`, creating a
    nearly-invisible clear view in the upper-left corner of the screen. The
    `accessibilityFrame` is set to match
    the Feature Highlight's bounds, which presents the affordance to
    UIAccessibility in a manner similar to a Scrim.  This is necessary to
    avoid introducing accidental "reject" taps if the user pressed on the
    Feature Highlight itself.

Closes #9450