[TextFields] Use correct localization API.
We were using the `NSLocalizedStringWithDefaultValue` API which is supposed to be used with the `genstrings` tool. This updates us to use the correct API which is `localizedStringForKey:value:table:`.

For more context on the `genstrings` tool see: https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/LoadingResources/Strings/Strings.html.

Closes #5616

PiperOrigin-RevId: 292469399