[ButtonBar] Assign the UIBarButtonItem's tintColor to the button's text color.
Prior to this change, UIBarButtonItem's tintColor property would not change the title color of the corresponding button. This differs from the UIKit behavior, which does result in the title color being changed.

After this change, changing a UIBarButtonItem's tintColor will result in the corresponding button's title color being updated as well.

PiperOrigin-RevId: 293583337