Add Swift-friendlier names to TextControl text fields
Automatic Swift header generation worked for the most part. However, stateful setters with the words "LabelColor" in them seemed to not translate correctly. For example...

This method:

-setTrailingAssistiveLabelColor:forState:

Would end up looking like this:

func setTrailingAssistiveLabel(_ trailingAssistiveLabelColor: UIColor, for state: MDCTextControlState)

Which makes it look like it's setting a label, not a label color.

Now those methods look like this:

func setTrailingAssistiveLabelColor(_ trailingAssistiveLabelColor: UIColor, for state: MDCTextControlState)

So when you call it it looks like this:

textField.setTrailingAssistiveLabelColor(.blue, for: .normal)

PiperOrigin-RevId: 293805941