[ButtonBar] Do not allow implicit downcasting of item views.
Prior to this change, it was possible to implicitly downcast item views to UIButton or MDCButton, potentially causing crashes by treating custom views incorrectly as a button and invoking UIButton selectors on it.

This change removes the `__kindof` implicit casting for the _buttonViews ivar so that the buttonViews need to be explicitly casted to a button type.

Closes https://github.com/material-components/material-components-ios/issues/9760