[FlexibleHeader] Subtract minimumHeaderViewHeight from fhv_accumulatorMax.
Problem statement: Attached banner on MDCFlexibleHeaderView are not intractable. Gestures on attached banner are not received, but rather goes through to view underneath the MDCFlexibleHeaderView.
Cause and Fix: MDCFlexibleHeaderView, after shifted to the top, with minimumHeaderViewHeight enabled, is shorter than intended. This was because fhv_accumulatorMax did not take minimumHeaderViewHeight into calculation. This corrects header view height after shifting.