[Chips] MDCChipCollectionViewCell to correctly support increased touch target (#9816)
Currently, the Chips component uses the `hitAreaInsets` API to allow an increase of the touch target of the view. However, when using MDCChipCollectionViewCell, setting the `hitAreaInsets` does not work, because there is no pointInside method to increase the tap target for the cell itself, which is responsible for the touches. This PR adds the same logic to the chip cell, borrowing the value of the hitAreaInsets from the chip view.

Tested in catalog with iOS 13 and iPhone Max Pro simulator and added unit tests.

Closes #9276