[BottomSheet] Anchor movement of the bottom sheet along the y axis.
Prior to this change, it was possible to make the bottom sheet "wobble" on the x-axis if you threw it just right, or if you adjusted the width of the application in a split-screen environment.

This bug was due to the fact that bottom sheet is implemented with UIDynamics and it was using a stock UIAttachmentBehavior to animate the bottom sheet to its destination. UIAttachmentBehavior supports both x and y movement out of the box, but bottom sheet appears to have historically attempted to mitigate this fact by resetting the x velocity to 0 in several places. For example:

https://github.com/material-components/material-components-ios/blob/a42a84cca9e9254ac56da2d4aed0ac03474d335e/components/BottomSheet/src/private/MDCDraggableView.m#L63-L64

and

https://github.com/material-components/material-components-ios/blob/a42a84cca9e9254ac56da2d4aed0ac03474d335e/components/BottomSheet/src/private/MDCDraggableView.m#L87-L88

After this change, the bottom sheet's allowed range of movement will only be along the y axis. This is now enforced through a custom action block that explicitly resets the x position of the behavior's target to _targetPoint.x.

TODO: Waiting on bug mirror to pick up the internal issue.