[Tabs] Add MDCTabBarViewLayoutStyleScrollableCentered (#9849)
This is a PR based off of [cl/298786100](http://cl/298786100). I was skeptical about the introduction of the new MDCTabViewAlignment enum in that CL, so I asked Rob (the author of the component) what he thought, and he suggested just extending the existing enum. I wanted to make sure that approach worked before I told the internal contributor about it, so I started tinkering with it, and eventually it just turned into this PR. If we want, we can just merge this PR instead of exporting that CL. I did take some of the code from that CL--this PR definitely would've taken longer to prepare without that code--so I'm wondering if this could still count as a client contribution?

Note: This PR changes the NSInteger values of some of the existing enum cases. I checked internally and none of the few people using this appear to be referring to the old values by their raw values, so I'm thinking this won't be a problem. If someone feels like it will be, I can add the new case to the bottom.

Note: The internal client is happy with this fix.

Here's a before gif:
![before-tabbar](https://user-images.githubusercontent.com/8020010/76020242-0d361980-5ef1-11ea-91b3-8099d2bc06af.gif)

Here's an after gif:
![after-tabbar](https://user-images.githubusercontent.com/8020010/76020231-09a29280-5ef1-11ea-8ec0-532a3f1ceea4.gif)

Closes #9422.