[FlexibleHeader] Fix bug where hairline configuration would not persist.
Prior to this change, the FlexibleHeader's hairline APIs were not backed by storage until the FlexibleHeader's view was loaded. This meant that settings on these properties would effectively get lost.

After this change, the FlexibleHeader initializes the hairline storage immediately on initialization so that the public hairline properties are consistently backed by storage.

Closes https://github.com/material-components/material-components-ios/issues/9863