This closes b/151091406, which was being caused because the presenting view controller is used to determine the frame of the bottom drawer's scroll view, which caused issues for situations in which the bottom drawer was presented from a modal view controller in the page or form sheet style.
This resolves said issue by adding an adjustment for the difference if applicable.

This change is necessary to fully support changes in modal presentation style as of iOS 13 + Xcode 11, which is important for clients because building with SDK 13/Xcode 11 is required for app store submission starting April 2020.

I verified this by testing the solution as it applies to our internal related bug, b/151091406, as well as by verifying this solution in the example project attached to the original reporting of this bug, b/148977218, and have added screencasts of this below (also attached to original bug for easier viewing).

As a sanity check, I also applied these changes to the catalog application and verified that it does not break any of the examples for this component.

PiperOrigin-RevId: 301395392