[Dialogs] Fix layout of configuration snapshot test.
Sizing dialogs snapshots to correct size in order to better simulate sizing of dialogs on a device.
The previous implementation used `preferredContentSize` which was not updated with the actual size of the snapshot after the dialog has been configured, causing most snapshot to be sized too wide.

This fix is implemented using the `calculatePreferredContentSizeForBounds:` API, which recalculates the dialog height based on the actual snapshot width, resulting in accurate width & height.

PiperOrigin-RevId: 301607700