# Problem The current snapshot is inaccurate because there is not enough room to layout all the dialog's elements, so spaces between the elements are inaccurate for many snapshots.
# Solution
Applying `sizeAlertToFitContent` before taking a snapshot of all dialogs. This method uses `calculatePreferredContentSizeForBounds:` to let the dialog layout itself correctly, within the given width, and to determine the height it needs to display itself.    The new snapshots have a more realistic layout that is closer to how dialogs should be presented.

PiperOrigin-RevId: 301862389