[MDC/Chips] Update intrinsicContentSize when text, font, or image change.
Currently since we expose the `titleLabel` and `imageView` of the `MDCChipView` when properties that can result in changes in size occur clients need to call `invalidateIntrinsicContentSize` in order to get the chip to update it's layout. This should be handled on our end instead of requiring clients to force the chip to layout when these properties change we can do it for them in the component. This issue occurs when the client has set the `autoresizingMask` to `.flexibleWidth, .flexibHeight` and constrained the position of the chip as seen in the tests.

I added unit test here instead of snapshot test because `BOOL success = [view drawViewHierarchyInRect:view.bounds afterScreenUpdates:YES];` in the snapshot test forces the chip to relay itself out causing the test to look like they are passing when they really aren't.

An alternative approach here would be to subclass UIImageView and UILabel and when the respective properties are set we can `invalidateIntrinsicContentSize` on the chip but I felt this case was better suited for KVO so we don't have to maintain these subclasses.

PiperOrigin-RevId: 302049886