Fix SnackBar crash when snackbarView is nil
These changes ensure that the CABasicAnimations involved in sliding snack bars on and off the screen are non-nil before they're added to an array.

I tested it by commenting out the following snackbarView-related code here to ensure that the snackbarView would be nil:

https://github.com/material-components/material-components-
ios/blob/develop/components/Snackbar/src/MDCSnackbarManager.m#L209-L230)

Without the changes in this CL it crashed, which was expected. With the changes, it did not crash. However, the snack bar did not show. Future changes should provide better error handling to help elucidate the circumstances in which the snackbarView. Improved error handling might entail adding NSError parameters to the completion blocks, for example. In the immediate term, however, my feeling is that we should just fix the crash.

PiperOrigin-RevId: 302490569