Only use self.snackbarView for dismissing snackbars, not presenting them
This CL does a nil check on snackbarView instead of self.snackbarView when presenting a snackbar. The method that displays snackbars is supposed to use an injected snackbarView instead of self.snackbarView, but the method that dismisses them is supposed to use self.snackbarView.

I broke this in cl/302513077. That CL adds additional nil checks, but it does so on self.snackbarView for both the display method and the dismissal method.

PiperOrigin-RevId: 302518117