[Dialogs] snapshot test for custom insets.
This CL adds a baseline snapshot for how the current dialog works.

The original layout of the Dialogs components used private insets that had all the same value, so many of the insets were used interchangeably. After creating separate insets for different sections, these issues because visible.
The goal of this CL is to demonstrate the issue of accessory view alignment (which is affected by content, title and actions insets). This issue is fixed in a followup cl.

PiperOrigin-RevId: 302970679